[Refactor] service 계층 구조화
