+++ src/interfaces/api/auth_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, Cookie, Response
 
 from src.interfaces.dependencies import get_auth_service
-from src.service.auth_service import AuthService
+from src.service.user.auth_service import AuthService
 from src.interfaces.schema.auth_schema import AuthResponse
 
 router = APIRouter(
