+++ src/interfaces/api/user_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends
 from src.interfaces.dependencies import get_user_service
 from src.interfaces.schema.user_schema import UuidResponse
-from src.service.user_service import UserService
+from src.service.user.user_service import UserService
 
 router = APIRouter(
     prefix="/api/user",
